fix(otp): cast digits/period to int in SELECT queries
Columns digits and period are TEXT in the actual database (despite
migration declaring INTEGER). pgx.RowToStructByName cannot scan
text into *int without an explicit cast. Add ::int to all SELECT
queries so the Go model (int fields) can scan correctly.

+++ services/otp/internal/repository/otp_repo.go
@@ -22,7 +22,7 @@ func NewOtpRepo(pool *pgxpool.Pool) *OtpRepo {
 }
 
 func (r *OtpRepo) FindByUserID(ctx context.Context, userID string) ([]model.Otp, error) {
-	rows, err := r.pool.Query(ctx, `SELECT id, user_id, email, secret, expires_at, type, issuer, digits, period, algorithm, valid FROM otp WHERE user_id = $1 ORDER BY id`, userID)
+	rows, err := r.pool.Query(ctx, `SELECT id, user_id, email, secret, expires_at, type, issuer, digits::int, period::int, algorithm, valid FROM otp WHERE user_id = $1 ORDER BY id`, userID)
 	if err != nil {
 		return nil, fmt.Errorf("query: %w", err)
 	}
@@ -31,7 +31,7 @@ func (r *OtpRepo) FindByUserID(ctx context.Context, userID string) ([]model.Otp,
 }
 
 func (r *OtpRepo) FindByID(ctx context.Context, id int64) (*model.Otp, error) {
-	rows, err := r.pool.Query(ctx, `SELECT id, user_id, email, secret, expires_at, type, issuer, digits, period, algorithm, valid FROM otp WHERE id = $1`, id)
+	rows, err := r.pool.Query(ctx, `SELECT id, user_id, email, secret, expires_at, type, issuer, digits::int, period::int, algorithm, valid FROM otp WHERE id = $1`, id)
 	if err != nil {
 		return nil, fmt.Errorf("query: %w", err)
 	}
@@ -72,7 +72,7 @@ func (r *OtpRepo) Update(ctx context.Context, o *model.Otp) error {
 }
 
 func (r *OtpRepo) FindByUserIDAndEmail(ctx context.Context, userID, email string) (*model.Otp, error) {
-	rows, err := r.pool.Query(ctx, `SELECT id, user_id, email, secret, expires_at, type, issuer, digits, period, algorithm, valid FROM otp WHERE user_id = $1 AND email = $2`, userID, email)
+	rows, err := r.pool.Query(ctx, `SELECT id, user_id, email, secret, expires_at, type, issuer, digits::int, period::int, algorithm, valid FROM otp WHERE user_id = $1 AND email = $2`, userID, email)
 	if err != nil {
 		return nil, fmt.Errorf("query: %w", err)
 	}
